Fix regression introduced by #156
Block indent should occur if there is a newline after the index form.
Before #156 the indent only occurred if there was a newline and then a
non-blank form. After #156 the indent occurred if there was a newline
and a non-black form, OR if the current line was empty. The latter
condition was overly broad; it did not check to see if the index
position had been filled, causing a regression.

Closes #159 and #160.