Add CLI option to parallelize file processing
The files processed by the 'fix' and 'check' commands can be processed
in an arbitrary order and each file can be processed independently, so
the whole process can be parallelized.

To make sure threads aren't writing to stdout/stderr at the same time,
the result for each file is captured and then printed in order once
every file is done. The 'check' command already did this via
print-file-status, the 'fix' command now uses the same mechanism.