Refactor indentation rule matching
Previously the indentation rules were sorted in the beginning and each
time an indentation decision had to be made, all rules were checked
consecutively until one matched.

This has a negative performance impact, as this check has to be
performed roughly once per newline in the file that is being processed.
It is in particular expensive because several low-level predicates have
to be re-used a lot on the same arguments for each indentation rule,
even tho those results never change.

Analysis showed that indent-matches?, form-symbol, token-value, token?
were the functions with the highest invocation count if indentation is
turned on, on an order of magnitude of about (number of rules) *
(number of lines). The default indentation rules alone add up to 81
symbols already.

This can be improved with another approach:
1. build look up maps for the indentation rule symbols at various
   depths, as required by the given ruleset
2. generate lazy maps for parent form symbols of a zloc, their
   qualified/unqualified versions, and a resulting set of relevant
   symbols to check at every relevant depth, which will only compute
   once per zloc and can then be accessed repeatedly with nth for any
   depth required
3. whenever an indentation decision has to be made all rules that can
   apply for the given zloc are enumerated based on the prepared maps,
   then sorted to guarantee the same order cljfmt favoured before, and
   finally the respective indentation functions are applied until an
   indentation is found

Performance:
It looks like JIT is very good at ironing over the re-used code paths.
It also is able to parallelize execution despite the code not using
concurrency itself.

For the uberjar I measured 2.4s user CPU time and 0.85s real time
before the change for a ~1.1K lines preformatted file. After the
change the result was 2.1s user CPU and 0.75s real time, so a 10-15%
speed improvement.
For a preformatted directory of about 260 files, totalling ~33K lines,
before the change it took 8.4s user CPU time and 4s real time, whereas
the change resulted in 7s user CPU time and 2.9s real time, a speed
improvement of 20-40%.

The bigger impact comes with a native GraalVM image, where I measured
0.29s / 0.33s for a single file and 6.2 / 6.3 for the directory before
the change.
And after the change I get 0.21 / 0.24 and 2.71 / 2.77 respectively.
An improvement of 35-40% for single files and ~220% for the directory.