Add support for zprint/cursive indentation
Add a new configuration option, :function-arguments-indentation, that
can be one of: :community, :cursive or :zprint. This controls the
behavior described in: https://guide.clojure.style/#one-space-indent

When set to :community (the default) it follows the Clojure style guide.
When set to :cursive or :zprint, it follows the conventions used by
those tools.

Also add a --function-arguments-indentation option to update this at the
command line.