Optionally indent "line comments" (set off by ;;)
Whole-line comments set off by exactly 2 semicolons (;;) are indented
to align with the program code, when option :indent-line-comments?
is true. The command-line flag is --indent-line-comments.

Batsov's Clojure Style Guide recommends, "Write comments on a
particular fragment of code before that fragment and aligned with it,
using two semicolons."  The Clojure Style Guide refers to them as
"line comments".

The :indent-line-comments? option is false by default, to avoid
changing the results of existing applictions of cljfmt and in
recognition that commenting conventions might not be universal.

Resolves: #222
See also: #112